+++ src/renderer/src/components/common/Slider.tsx
@@ -1,4 +1,4 @@
-import { ChangeEvent, ReactNode } from 'react';
+import { ChangeEvent, CSSProperties, ReactNode } from 'react';
 
 type Props = {
   label?: ReactNode;
@@ -31,13 +31,13 @@ export const Slider = ({ label, min, max, step = 1, value, onChange, className =
         step={step}
         value={value}
         onChange={handleChange}
-        className="w-full h-2 bg-neutral-700 rounded-lg appearance-none cursor-pointer accent-neutral-200 transition-colors bg-slider-track"
+        className="w-full h-2 bg-neutral-700 rounded-lg appearance-none cursor-pointer accent-neutral-200 transition-colors bg-slider-track mt-4 mb-[9px]"
         style={
           {
             '--slider-percentage': `${((value - min) / (max - min)) * 100}%`,
             '--slider-filled-color': 'var(--tw-color-neutral-500)',
             '--slider-empty-color': 'var(--tw-color-neutral-700)',
-          } as React.CSSProperties
+          } as CSSProperties
         }
       />
     </div>