+++ src/renderer/src/components/settings/providers/index.ts
@@ -1,0 +1,5 @@
+export * from './OpenAiParameters';
+export * from './AnthropicParameters';
+export * from './GeminiParameters';
+export * from './BedrockParameters';
+export * from './ModelSelect';